initializeEditorState on LexicalComposer
