+++ packages/lexical-markdown/README.md
@@ -22,10 +22,10 @@ editor.update(() => {
 
 It can also be used for initializing editor's state from markdown string. Here's an example with react `<RichTextPlugin>`
 ```jsx
-<LexicalComposer>
-  <RichTextPlugin initialEditorState={() => {
+<LexicalComposer initialEditorState={() => {
     $convertFromMarkdownString(markdown, TRANSFORMERS);
-  }} />
+  }}>
+  <RichTextPlugin />
 </LexicalComposer>
 ```
