+++ packages/lexical-playground/src/nodes/ImageNode.tsx
@@ -288,7 +288,6 @@ function ImageComponent({
                     Enter a caption...
                   </Placeholder>
                 }
-                initialEditorState={null}
               />
               {showNestedEditorTreeView === true ? <TreeViewPlugin /> : null}
             </LexicalNestedComposer>
