+++ packages/lexical-react/src/LexicalRichTextPlugin.tsx
@@ -9,10 +9,15 @@
 import {useLexicalComposerContext} from '@lexical/react/LexicalComposerContext';
 import {InitialEditorStateType} from '@lexical/rich-text';
 import * as React from 'react';
+import warnOnlyOnce from 'shared/warnOnlyOnce';
 
 import {useCanShowPlaceholder} from './shared/useCanShowPlaceholder';
 import {useDecorators} from './shared/useDecorators';
 import {useRichTextSetup} from './shared/useRichTextSetup';
+
+const deprecatedInitialEditorStateWarning = warnOnlyOnce(
+  'initialEditorState on RichTextPlugin is deprecated and will be removed soon. Use LexicalComposer initialEditorState instead.',
+);
 
 export function RichTextPlugin({
   contentEditable,
@@ -20,16 +25,17 @@ export function RichTextPlugin({
   initialEditorState,
 }: Readonly<{
   contentEditable: JSX.Element;
+  // TODO Remove in 0.4
   initialEditorState?: InitialEditorStateType;
   placeholder: JSX.Element | string;
 }>): JSX.Element {
+  if (initialEditorState !== undefined) {
+    deprecatedInitialEditorStateWarning();
+  }
   const [editor] = useLexicalComposerContext();
-
   const showPlaceholder = useCanShowPlaceholder(editor);
-
-  useRichTextSetup(editor, initialEditorState);
-
   const decorators = useDecorators(editor);
+  useRichTextSetup(editor, initialEditorState);
 
   return (
     <>
