+++ packages/lexical-react/src/__tests__/unit/PlainRichTextPlugin.test.tsx
@@ -67,6 +67,7 @@ describe('LexicalNodeHelpers tests', () => {
         return (
           <LexicalComposer
             initialConfig={{
+              editorState: $initialEditorState,
               namespace: '',
               nodes:
                 plugin === 'PlainTextPlugin'
@@ -95,13 +96,11 @@ describe('LexicalNodeHelpers tests', () => {
             {plugin === 'PlainTextPlugin' ? (
               <PlainTextPlugin
                 contentEditable={<ContentEditable />}
-                initialEditorState={$initialEditorState}
                 placeholder=""
               />
             ) : (
               <RichTextPlugin
                 contentEditable={<ContentEditable />}
-                initialEditorState={$initialEditorState}
                 placeholder=""
               />
             )}
@@ -134,6 +133,7 @@ describe('LexicalNodeHelpers tests', () => {
         return (
           <LexicalComposer
             initialConfig={{
+              editorState: initialEditorStateJson,
               namespace: '',
               nodes:
                 plugin === 'PlainTextPlugin'
@@ -162,13 +162,11 @@ describe('LexicalNodeHelpers tests', () => {
             {plugin === 'PlainTextPlugin' ? (
               <PlainTextPlugin
                 contentEditable={<ContentEditable />}
-                initialEditorState={initialEditorStateJson}
                 placeholder=""
               />
             ) : (
               <RichTextPlugin
                 contentEditable={<ContentEditable />}
-                initialEditorState={initialEditorStateJson}
                 placeholder=""
               />
             )}
